requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.3.0

# Gemini & LangChain
google-generativeai>=0.8.0
langchain>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector Store
chromadb>=0.5.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0
pdfplumber>=0.11.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20

# Utilities
aiofiles>=24.1.0
tenacity>=9.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
